pyproject.toml
[project]
name = "sumtech-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.119.0",
    "openai>=2.6.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
